% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstyle.R
\name{style_file}
\alias{style_file}
\title{style a file}
\usage{
style_file(path)
}
\value{
nothing modifies file as side effect.
}
\description{
Style a file
}
\details{
style a file with my variant of styler::tidyverse_style()
}
\author{
Miles McBain
}
